demoapp/requirements.txt
django
djangorestframework
psycopg2-binary
requests
python-dotenv
